+++ Changelog.md
@@ -2,6 +2,7 @@ AppleALC Changelog
 ==================
 #### v1.5.5
 - Fixed kext loading issues on 10.8 and similar operating system
+- Added a requirement to inject `alc-verbs` or use `alc-verbs=1` argument for custom verbs
 
 #### v1.5.4
 - Improved Ice Lake controller patches by fewtarius